# frontend/openapi.json
{
	"components": {
		"schemas": {
			"APIErrorBody": {
				"properties": {
					"code": {
						"title": "Code",
						"type": "string"
					},
					"message": {
						"title": "Message",
						"type": "string"
					},
					"request_id": {
						"title": "Request Id",
						"type": "string"
					},
					"retryable": {
						"default": false,
						"title": "Retryable",
						"type": "boolean"
					}
				},
				"required": ["code", "message", "request_id"],
				"title": "APIErrorBody",
				"type": "object"
			},
...
			"ChannelCreate": {
				"description": "Schema for adding a new channel. The user provides the handle\nand optionally which folder to place it in.",
				"properties": {
					"folder_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Folder Id"
					},
					"handle": {
						"title": "Handle",
						"type": "string"
					}
				},
				"required": ["handle"],
				"title": "ChannelCreate",
				"type": "object"
			},
			"ChannelOut": {
				"description": "Schema for returning a channel from the API.",
				"properties": {
					"created_at": {
						"format": "date-time",
						"title": "Created At",
						"type": "string"
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description"
					},
					"folder_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Folder Id"
					},
					"handle": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Handle"
					},
					"id": {
						"title": "Id",
						"type": "string"
					},
					"is_favorited": {
						"title": "Is Favorited",
						"type": "boolean"
					},
					"last_updated": {
						"format": "date-time",
						"title": "Last Updated",
						"type": "string"
					},
					"thumbnail_url": {
						"anyOf": [
							{
								"format": "uri",
								"maxLength": 2083,
								"minLength": 1,
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Thumbnail Url"
					},
					"title": {
						"title": "Title",
						"type": "string"
					},
					"total_videos": {
						"default": 0,
						"title": "Total Videos",
						"type": "integer"
					}
				},
				"required": [
					"id",
					"title",
					"handle",
					"description",
					"thumbnail_url",
					"is_favorited",
					"folder_id",
					"created_at",
					"last_updated"
				],
				"title": "ChannelOut",
				"type": "object"
			},
...
			"PaginatedResponse_ChannelOut_": {
				"properties": {
					"has_more": {
						"title": "Has More",
						"type": "boolean"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ChannelOut"
						},
						"title": "Items",
						"type": "array"
					},
					"limit": {
						"title": "Limit",
						"type": "integer"
					},
					"offset": {
						"title": "Offset",
						"type": "integer"
					},
					"total": {
						"title": "Total",
						"type": "integer"
					}
				},
				"required": ["total", "items", "limit", "offset", "has_more"],
				"title": "PaginatedResponse[ChannelOut]",
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "auth/jwt/login"
					}
				},
				"type": "oauth2"
			}
		}
	},
	"info": {
		"license": {
			"identifier": "GPL-3.0-only",
			"name": "GPL-3.0-only"
		},
		"summary": "Backend API for the ChooseYourTube distraction-free YouTube reader.",
		"title": "ChooseYourTube API",
		"version": "0.1.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/channels/": {
			"delete": {
				"description": "Deletes ALL channels from the database.\nThis is a destructive operation intended for testing.\n\nRequires confirmation parameter: ?confirm=DELETE_ALL_CHANNELS",
				"operationId": "delete_all_channels_channels__delete",
				"parameters": [
					{
						"description": "Must be exactly 'DELETE_ALL_CHANNELS' to confirm this destructive operation",
						"in": "query",
						"name": "confirm",
						"required": true,
						"schema": {
							"description": "Must be exactly 'DELETE_ALL_CHANNELS' to confirm this destructive operation",
							"title": "Confirm",
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Delete All Channels",
				"tags": ["Channels"]
			},
			"get": {
				"description": "Retrieves a list of all channels with optional filtering and pagination.\n\nFilters:\n- is_favorited: Show only favorited channels\n- folder_id: Show only channels in a specific folder (use 0 for channels with no folder)\n- tag_id: Show only channels with a specific tag\n\nPagination:\n- limit: Number of items per page (default: 50, max: 200)\n- offset: Number of items to skip",
				"operationId": "list_channels_channels__get",
				"parameters": [
					{
						"description": "Filter by favorited status",
						"in": "query",
						"name": "is_favorited",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "boolean"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by favorited status",
							"title": "Is Favorited"
						}
					},
					{
						"description": "Filter by folder ID (use 0 for root/no folder)",
						"in": "query",
						"name": "folder_id",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by folder ID (use 0 for root/no folder)",
							"title": "Folder Id"
						}
					},
					{
						"description": "Filter by tag ID",
						"in": "query",
						"name": "tag_id",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by tag ID",
							"title": "Tag Id"
						}
					},
					{
						"description": "Number of items per page",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"description": "Number of items per page",
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"description": "Number of items to skip",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"description": "Number of items to skip",
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_ChannelOut_"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "List Channels",
				"tags": ["Channels"]
			},
			"post": {
				"description": "Creates a new YouTube channel in the application.\n- Fetches channel details from the YouTube Data API.\n- Stores the channel information in the database.",
				"operationId": "create_channel_channels__post",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChannelCreate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChannelOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Create Channel",
				"tags": ["Channels"]
			}
		},
		"/channels/{channel_id}": {
			"delete": {
				"description": "Deletes a single channel by its YouTube Channel ID.\nAll associated videos will also be deleted.",
				"operationId": "delete_channel_channels__channel_id__delete",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Delete Channel",
				"tags": ["Channels"]
			},
			"get": {
				"description": "Retrieves a single channel by its YouTube Channel ID.",
				"operationId": "get_channel_by_id_channels__channel_id__get",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChannelOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Get Channel By Id",
				"tags": ["Channels"]
			},
			"patch": {
				"operationId": "update_channel_channels__channel_id__patch",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChannelUpdate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChannelOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Update Channel",
				"tags": ["Channels"]
			}
		},
		"/channels/{channel_id}/playlists": {
			"get": {
				"operationId": "list_channel_playlists_channels__channel_id__playlists_get",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					},
					{
						"description": "Include inactive synced playlists",
						"in": "query",
						"name": "include_inactive",
						"required": false,
						"schema": {
							"default": false,
							"description": "Include inactive synced playlists",
							"title": "Include Inactive",
							"type": "boolean"
						}
					},
					{
						"description": "Number of items per page",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"description": "Number of items per page",
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"description": "Number of items to skip",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"description": "Number of items to skip",
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_ChannelPlaylistOut_"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "List Channel Playlists",
				"tags": ["Channels"]
			}
		},
		"/channels/{channel_id}/playlists/refresh": {
			"post": {
				"operationId": "refresh_channel_playlists_channels__channel_id__playlists_refresh_post",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"202": {
						"content": {
							"application/json": {
								"schema": {}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Refresh Channel Playlists",
				"tags": ["Channels"]
			}
		},
		"/channels/{channel_id}/refresh": {
			"post": {
				"description": "Refresh the given YouTube channel to add and update the first 50 videos in its uploads playlist",
				"operationId": "refresh_channel_channels__channel_id__refresh_post",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChannelOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Refresh Channel",
				"tags": ["Channels"]
			}
		},
...
		"/tags/{tag_id}/channels": {
			"get": {
				"description": "Get all channels associated with a tag.",
				"operationId": "get_channels_for_tag_tags__tag_id__channels_get",
				"parameters": [
					{
						"in": "path",
						"name": "tag_id",
						"required": true,
						"schema": {
							"title": "Tag Id",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChannelOut"
									},
									"title": "Response Get Channels For Tag Tags  Tag Id  Channels Get",
									"type": "array"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Get Channels For Tag",
				"tags": ["Tags"]
			}
		},
...
	}
}